Added context to knife bootstrap instructions
It is not obvious from the the knife bootstrap documentation that you
knife bootstrap will SSH onto the target machine. I have expanded the first
paragraph to explicitly state that and provide enough context for the rest
of the documentation. Hopefully, people coming in cold will not try to
run knife bootstrap on the node machine like I did.