change the install workstation link
There are 2 install workstation page in the wiki (pretty confusing)
http://docs.opscode.com/chef/install_workstation.html
http://docs.opscode.com/install_workstation.html

For me, the second one is better, that's why, I changed the reference
there.

Can you explain me the purpose of both please? (if there is a different
purpose, I not, I can take care of removing references to the 1st one)