Adds a way to handle the nav
This creates a way to override a block inside the main left nav. The
idea here being that the default/standard nav will live in the theme
with the ability to override on a per docset level. For example, on the
delivery doc sets we are going to have a custom nav from the main site,
but for ohai we may just leave the default.