FIX to the user resource documentation
The statement 'create a random user' makes me think this would generate a user at random. This in fact is a user named 'random' that is being created.

The other example of creating a system user with a variable had a number of issues with the syntax. A node variable cannot be used in a single-quoted string like that. Also it does not make sense to create a user with the value of the group attribute. The `user_home` was not creating a directory for home when it probably should be building a path with the user name. The use of the variable `user_hame` also cannot be used with single-quotes around it. Ruby treats that as a the string literally containing the value 'user_home'.