Don't allow for caching at all
A while back we disabled caching because it was causing problems. This
takes that even further by creating the ebs volume on the fly so that we
don't wait for the `rm -rf` to complete. This also addresses an issue
where a build would roll directly in and fail because the shared ebs
volume was still mounted by test kitchen in another job.